% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{acc}
\alias{acc}
\title{Compute prediction accuracy for a quantitative or binary trait}
\usage{
acc(yobs = NULL, ypred = NULL, typeoftrait = "quantitative")
}
\arguments{
\item{yobs}{is a vector of observed phenotypes}

\item{ypred}{is a vector of predicted phenotypes}

\item{typeoftrait}{is a character with possible values "binary" or "quantitative" (default)}
}
\description{
Compute prediction accuracy for a quantitative or binary trait
}
