% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{auc}
\alias{auc}
\title{Compute AUC}
\usage{
auc(yobs = NULL, ypred = NULL)
}
\arguments{
\item{yobs}{is a vector of observed phenotypes}

\item{ypred}{is a vector of predicted phenotypes}
}
\description{
Compute Area Under the Curve (AUC)
}
\keyword{internal}
