% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgfm.R
\name{bgfm}
\alias{bgfm}
\title{Genomic Feature Model analyses implemented using Bayesian Methods}
\usage{
bgfm(y = NULL, g = NULL, nsamp = 50, nburn = 10, nsave = 10000,
  tol = 0.001)
}
\arguments{
\item{y}{is a matrix of phenotypes}

\item{g}{is a list of G matrices (g$G), prior (co)variances (g$sigma), prior degrees of freedom (g$sigma),  prior distribution (g$ds)}

\item{nsamp}{is the number of samples after burnin}

\item{nburn}{is the number of burnin samples}

\item{nsave}{is the number of samples to save}

\item{tol}{is the tolerance}
}
\value{
Returns a list including 
\item{sigmas}{posterior samples variance components} 
\item{alphas}{posterior samples alphas}
\item{mus}{posterior samples mu}
\item{logCPO}{log CPOs}
\item{g}{g is a list of G matrices (g$G), prior (co)variances (g$sigma), prior degrees of freedom (g$sigma),  prior distribution (g$ds)}
}
\description{
Bayesian Genomic Feature models implemented using Bayesian Methods. 
Multiple features and multiple traits models can be fitted.
Different genetic models (e.g. additive, dominance, gene by gene and gene by environment interactions) can be specified.
}
\details{
The models are implemented using empirical Bayesian methods. The hyperparameters of the dispersion parameters of the Bayesian model can
be obtained from prior information or estimated by maximum likelihood, and conditional on these, the model is fitted using
Markov chain Monte Carlo. Furthermore, a spectral decomposition of genomic feature relationship matrices plays an 
important computational role in the Markov chain Monte Carlo strategy implemented.
}
\author{
Peter Sørensen
}
