% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_score.R
\name{computeROC}
\alias{computeROC}
\title{Compute Receiver Operating Curve statistics}
\usage{
computeROC(yobs = NULL, ypred = NULL)
}
\arguments{
\item{yobs}{vector of observed phenotype}

\item{ypred}{vector of predicted phenotype}
}
\description{
Compute ROC
}
\author{
Palle Duun Rohde
}
\keyword{internal}
