% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qgg.R
\name{cvat}
\alias{cvat}
\title{Genetic marker set tests based on the covariance test}
\usage{
cvat(s = NULL, g = NULL, W = NULL, sets = NULL, nperm = 100)
}
\arguments{
\item{s}{vector (or list) of single marker effects obtained from a linear mixed model fit (GBLUP of GFBLUP)}

\item{g}{vector (or list) of genetic effects obtained from a linear mixed model fit (GBLUP of GFBLUP)}

\item{W}{matrix of centered and scaled genotypes (nxm)}

\item{sets}{list of marker sets corresponding to columns in W}

\item{nperm}{number of permutations}
}
\value{
Returns a dataframe including
\item{setT}{covariance test statistics}
\item{nset}{number of markers in the set }
\item{p}{p-value }
}
\description{
Genetic marker set tests based on the covariance statistics for a set of genetic markers
}
\details{
The covariance test statistic is derived from a GBLUP( or GFBLUP) model fit. It is a measure of covariance between the total genomic effect for all markers
and the genomic effect for the genetic markers in the genomic feature. It also relates to the explained sums of
squares for the genetic markers.
The distribution of this test statistic under the null hypothesis is difficult to describe in terms of exact or approximate
distributions, and an empirical distribution is required.
}
\examples{
# Simulate data
W <- matrix(rnorm(2000000),ncol=10000)
  colnames(W) <- as.character(1:ncol(W))
y <- rowSums(W[,1:10]) + rowSums(W[,1001:1010]) + rnorm(nrow(W))

# REML analyses
data <- data.frame(y=y,mu=1)
fm <- y ~ mu
fit <- gfm(fm=fm,W=W,sets=list(colnames(W)),data=data)

# cvat set test
sets <- list(A=as.character(1:100),B=as.character(101:1000),C=as.character(1001:5000),D=as.character(5001:10000))
res <- cvat(s=fit$s,W=W,sets=sets,nperm=100)
res
}
\author{
Peter Sørensen
}
\references{
Genomic BLUP Derived Set Tests Identify Genetic Variants Associated with Schizophrenia in Functionally Associated Biological Processes. Under review, Genetics (2015). Rohde PD, Demontis D, The GEMS Group, Børglum AD, Sørensen P.
}

