% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_grm.R
\name{eigGRM}
\alias{eigGRM}
\title{Eigen value decompostion of a genomic relationship matrix}
\usage{
eigGRM(GRM = NULL, GRMlist = NULL, method = "default", ncores = 1)
}
\arguments{
\item{GRM}{genomic relationship matrix}

\item{GRMlist}{list of information about GRM matrix stored on disk}

\item{method}{used for decompositon}

\item{ncores}{the number of cores used}
}
\description{
A function for obtaining eigen values and vectors of a GRM. Supports GRM stored in a 
binary file on disk
}
\examples{

# Simulate data
W <- matrix(rnorm(20000000), ncol = 10000)
	colnames(W) <- as.character(1:ncol(W))
	rownames(W) <- as.character(1:nrow(W))

# Compute G
GRM <- computeGRM(W = W)

eig <- eigGRM(GRM=GRM)
}
