% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_reml.R
\name{gblup}
\alias{gblup}
\title{Compute Genomic BLUP values}
\usage{
gblup(
  GRMlist = NULL,
  GRM = NULL,
  fit = NULL,
  ids = NULL,
  idsCLS = NULL,
  idsRWS = NULL
)
}
\arguments{
\item{GRMlist}{list providing information about GRM matrix stored in binary files on disk}

\item{GRM}{list of one or more genomic relationship matrices}

\item{fit}{list object output from greml function}

\item{ids}{vector of ids for which BLUP values is computed}

\item{idsCLS}{vector of column ids in GRM for which BLUP values is computed}

\item{idsRWS}{vector of row ids in GRM for which BLUP values is computed}
}
\description{
Compute Genomic BLUP values based on linear mixed model fit output from greml
}
\keyword{internal}
