% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_bayes.R
\name{getCredibleSets}
\alias{getCredibleSets}
\title{Get Credible Sets}
\usage{
getCredibleSets(Glist = NULL, fit = NULL, sets = NULL)
}
\arguments{
\item{Glist}{A list containing genomic data (e.g., `chr` and `bedfiles`).}

\item{fit}{A model object containing SNP statistics (`stat` with `rsids` and `dm` columns).}

\item{sets}{A list of marker sets (SNP identifiers).}
}
\value{
A list of credible sets for each marker set.
}
\description{
This function calculates credible sets for a list of marker sets using probabilistic measures.
}
