% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_grm.R
\name{getGRM}
\alias{getGRM}
\title{Extract elements from genomic relationship matrix (GRM) stored on disk}
\usage{
getGRM(
  GRMlist = NULL,
  ids = NULL,
  idsCLS = NULL,
  idsRWS = NULL,
  cls = NULL,
  rws = NULL
)
}
\arguments{
\item{GRMlist}{list providing information about GRM matrix stored in binary files on disk}

\item{ids}{vector of ids in GRM to be extracted}

\item{idsCLS}{vector of column ids in GRM to be extracted}

\item{idsRWS}{vector of row ids in GRM to be extracted}

\item{cls}{vector of columns in GRM to be extracted}

\item{rws}{vector of rows in GRM to be extracted}
}
\description{
Extract elements from genomic relationship matrix (GRM) (whole or subset) stored on disk.
}
\keyword{internal}
