% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_matrix.R
\name{getLDsets}
\alias{getLDsets}
\title{Get marker LD sets}
\usage{
getLDsets(Glist = NULL, chr = NULL, r2 = 0.5)
}
\arguments{
\item{Glist}{A list structure containing information about genotypes stored on disk.}

\item{chr}{A numeric value specifying the chromosome for which LD sets are to be extracted.}

\item{r2}{A numeric threshold, defaulting to 0.5, used for extracting LD sets.}
}
\description{
Extracts marker LD sets based on a sparse LD matrix stored in the Glist object.
}
\keyword{internal}
