% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_matrix.R
\name{getMarkers}
\alias{getMarkers}
\title{Retrieve marker rsids in a specified genome region.}
\usage{
getMarkers(Glist = NULL, chr = NULL, region = NULL)
}
\arguments{
\item{Glist}{A list structure with information about genotypes stored on disk.}

\item{chr}{A chromosome from which markers are extracted.}

\item{region}{A genome region (in base pairs) from which markers are extracted.}
}
\value{
A vector of rsids that fall within the specified region on the given chromosome.
}
\description{
Get marker rsids (reference SNP cluster IDs) from a specified genome region 
based on markers present in `Glist`.
}
\keyword{internal}
