% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_genotype_matrix.R
\name{getW}
\alias{getW}
\title{Extract elements from genotype matrix (W) stored on disk}
\usage{
getW(Glist = NULL, ids = NULL, rsids = NULL, rws = NULL,
  cls = NULL, impute = FALSE, scale = FALSE, allele = NULL)
}
\arguments{
\item{Glist}{only provided if task="summary" or task="sparseld"}

\item{ids}{vector of ids in W to be extracted}

\item{rsids}{vector of rsids in W to be extracted}

\item{rws}{vector of rows in W to be extracted}

\item{cls}{vector of columns in W to be extracted}

\item{impute}{logical if TRUE missing genotypes are set to its expected value (2*af where af is allele frequency)}

\item{scale}{logical if TRUE the genotype markers have been scale to mean zero and variance one}

\item{allele}{vector of alleles to be extracted}
}
\description{
Extract elements from genotype matrix W (whole or subset) stored on disk.
}
