% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qgg.R
\name{gfm}
\alias{gfm}
\title{Genomic Feature Model analyses implemented using Likelihood Methods}
\usage{
gfm(fm = NULL, weights = NULL, W = NULL, sets = NULL, K = NULL,
  data = NULL, validate = NULL, mkplots = TRUE)
}
\arguments{
\item{fm}{a formula with model statement for the fixed factors in the linear mixed model}

\item{weights}{a vector of weights for the residual variance}

\item{W}{a matrix centered and scaled genotypes or other types of molecular data}

\item{sets}{a list of marker sets corresponding to column names in W}

\item{K}{a list of correlation matrices}

\item{data}{a data frame containing the phenotypic observations and fixed factors specified in the model statements}

\item{validate}{a matrix or a list with the ids of validation sets corresponding to the rows in data}

\item{mkplots}{a logical indicating whether or not to make plots}
}
\value{
Returns results in a list structure including
\item{f}{matrix of predicted random effects (training set)}
\item{fv}{matrix of predicted random effects validation set}
\item{Vf}{covariance of predicted random effects}
\item{s}{single marker effects}
\item{vs}{variance of single marker effects}
\item{sigmas}{estimated variance components}
\item{pa}{predictive ability based on validation sets}
\item{mspe}{mean squared prediction error based on validation sets}
\item{ypred}{predicted phenotypes validation set}
\item{yobs}{observed phenotypes validation set}
\item{fit}{fit object from regress}
\item{validate}{matrix of validation sets as provided in arguments}
}
\description{
Genomic Feature Best Linear Unbiased Prediction models implemented using REML.
Multiple features and multiple traits models can be fitted.
Different genetic models (e.g. additive, dominance, gene by gene and gene by environment interactions) can be specified.
}
\details{
The models are implemented using restricted maximum likelihood methods.
Variance components estimated using REML and predictions are based on BLUP.
First and second derivatives of log-likehood can be obtained in addition to asymtotic standard deviation of parameter estimates.
Predicted random effects and single marker effects and statistics can be obtained.
Covariance decomposition procedures.
Cross validation procedures for assessment of prediction accuracy and model selection.
Currently depend on R library regress.
Plug-in avaliable for DMU. Future release may use customized REML procedures.
}
\examples{
# Simulate data
W <- matrix(rnorm(2000000),ncol=10000)
  colnames(W) <- as.character(1:ncol(W))
y <- rowSums(W[,1:10]) + rowSums(W[,1001:1010]) + rnorm(nrow(W))

# REML analyses and cross validation
data <- data.frame(y=y,mu=1)
fm <- y ~ mu

setsGB <- list(A=colnames(W))
setsGF <- list(C1=colnames(W)[1:1000],C2=colnames(W)[1001:2000],C2=colnames(W)[2000:10000])
setsGT <- list(C1=colnames(W)[1:10],C2=colnames(W)[1001:1010],C2=colnames(W)[1:10000])

n <- length(y)
fold <- 10
nsets <- 50
validate <- replicate(nsets,sample(1:n,as.integer(n/fold)))

fitGB <- gfm(fm=fm,W=W,sets=setsGB,data=data,validate=validate)
fitGF <- gfm(fm=fm,W=W,sets=setsGF,data=data,validate=validate)
fitGT <- gfm(fm=fm,W=W,sets=setsGT,data=data,validate=validate)


validClust <- list(C1=1:40,C2=41:77)

fitGB <- gfm(fm=fm,W=W,sets=setsGB,data=data,validate=validClust)
fitGF <- gfm(fm=fm,W=W,sets=setsGF,data=data,validate=validClust)
fitGT <- gfm(fm=fm,W=W,sets=setsGT,data=data,validate=validClust)


# REML analyses and variance decomposition

fitGB <- gfm(fm=fm,W=W,data=data)
covSets(W=W,g=fitGB$f,sets=setsGF)$cvf
covSets(W=W,g=fitGB$f,sets=setsGT)$cvf


fitGT <- gfm(fm=fm,W=W,sets=setsGT,data=data)
covSets(W=W,g=fitGT$f[,1],sets=setsGT)$cvf
covSets(W=W,g=fitGT$f[,2],sets=setsGT)$cvf

cvgf1 <- covSets(W=W,g=fitGT$f[,1],sets=setsGT, level2=TRUE)
cvgf2 <- covSets(W=W,g=fitGT$f[,2],sets=setsGT, level2=TRUE)

bars1 <- cvgf1$cvf[,2]
bars11 <- cvgf1$cvfSet[[1]][,5]
bars12 <- cvgf1$cvfSet[[2]][,5]

bars2 <- cvgf2$cvf[,2]
bars21 <- cvgf2$cvfSet[[1]][,5]
bars22 <- cvgf2$cvfSet[[2]][,5]

layout(matrix(1:6,ncol=2,byrow=TRUE))
barplot(bars1)
barplot(bars2)
barplot(bars11)
barplot(bars12)
barplot(bars21)
barplot(bars22)



# REML analyses
data <- data.frame(y=y,mu=1)
fm <- y ~ mu
fit <- gfm(fm=fm,W=W,sets=list(colnames(W)),data=data)

# REML analyses
pc1 <- rnorm(200)
pc2 <- rnorm(200)
f1 <- factor(rep(1:2,100))
fmf <- y~pc1+pc1+f1
fitGB <- gfm(fm=fmf,W=W,data=data.frame(y,pc1,pc1,f1),validate=validate)
fitGF <- gfm(fm=fmf,W=W,sets=setsGF,data=data.frame(y,pc1,pc1,f1),validate=validate)
fitGT <- gfm(fm=fmf,W=W,sets=setsGT,data=data.frame(y,pc1,pc1,f1),validate=validate)
}
\author{
Peter Sørensen
}
\references{
Mapping Variants to Gene Ontology Categories Improves Genomic Prediction for Quantitative Traits in Drosophila melanogaster. Under review Genetics (2016). Edwards SM, Sørensen IF, Sarup P, Mackay TF, Sørensen P.

Genomic BLUP Derived Set Tests Identify Genetic Variants Associated with Schizophrenia in Functionally Associated Biological Processes. Under review, Genetics (2015). Rohde PD, Demontis D, The GEMS Group, Børglum AD, Sørensen P.

Partitioning of genomic variance reveals biologic pathways associated with udder health and milk production traits in dairy cattle. GSE (2015) 47:60. Edwards SM, Thomsen B, Madsen P, Sørensen P.

Increased prediction accuracy using a genomic feature model including prior information on quantitative trait locus regions in purebred Danish Duroc pigs. BMC Genetics (2016) 17:11. Sarup P, Jensen J, OstersenT, Henryon M, Sørensen P.
}

