% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_bayes.R
\name{gmap}
\alias{gmap}
\title{Finemapping using Bayesian Linear Regression Models}
\usage{
gmap(
  Glist = NULL,
  stat = NULL,
  sets = NULL,
  models = NULL,
  rsids = NULL,
  ids = NULL,
  mask = NULL,
  lambda = NULL,
  vb = NULL,
  vg = NULL,
  ve = NULL,
  pi = 0.001,
  h2 = 0.5,
  nub = 4,
  nug = 4,
  nue = 4,
  ssb_prior = NULL,
  ssg_prior = NULL,
  sse_prior = NULL,
  vb_prior = NULL,
  vg_prior = NULL,
  ve_prior = NULL,
  updateB = TRUE,
  updateG = TRUE,
  updateE = TRUE,
  updatePi = TRUE,
  formatLD = "dense",
  checkLD = FALSE,
  shrinkLD = FALSE,
  shrinkCor = FALSE,
  pruneLD = FALSE,
  checkConvergence = FALSE,
  critVe = 3,
  critVg = 3.5,
  critVb = 3.5,
  critPi = 3,
  ntrial = 1,
  verbose = FALSE,
  eigen_threshold = 0.995,
  credible_set_r2 = 0.5,
  r2 = 0.05,
  nit = 5000,
  nburn = 500,
  nthin = 5,
  method = "bayesR",
  algorithm = "mcmc-eigen",
  seed = 10
)
}
\arguments{
\item{Glist}{A list containing information on genotypic data, including SNPs, chromosomes, positions, and optionally, LD matrices.}

\item{stat}{A data frame or list of summary statistics including effect sizes, standard errors, sample sizes, etc.}

\item{sets}{Optional list specifying sets of SNPs for mapping.}

\item{models}{Optional list of predefined models for Bayesian regression.}

\item{rsids}{Vector of SNP identifiers.}

\item{ids}{Vector of sample identifiers.}

\item{mask}{Logical matrix indicating SNPs to exclude from analysis.}

\item{lambda}{Vector of initial values for penalty parameters in the model.}

\item{vb}{Initial prior for the marker effect variance (default: NULL).}

\item{vg}{Initial prior for the genetic variance (default: NULL).}

\item{ve}{Initial prior for the residual variance (default: NULL).}

\item{pi}{Initial probability of inclusion (default: 0.001).}

\item{h2}{Heritability estimate (default: 0.5).}

\item{nub, nug, nue}{Degrees of freedom parameters for the priors of marker, genetic, and residual variances, respectively.}

\item{ssb_prior, ssg_prior, sse_prior}{Priors for the marker, genetic, and residual variances.}

\item{vb_prior, vg_prior, ve_prior}{Additional priors for marker, genetic, and residual variances (default: NULL).}

\item{updateB, updateG, updateE, updatePi}{Logical values specifying whether to update marker effects, genetic variance, residual variance, and inclusion probabilities, respectively.}

\item{formatLD}{Format of LD matrix ("dense" by default).}

\item{checkLD}{Logical, whether to check the LD matrix for inconsistencies (default: FALSE).}

\item{shrinkLD, shrinkCor}{Logical, whether to apply shrinkage to the LD or correlation matrices (default: FALSE).}

\item{pruneLD}{Logical, whether to prune LD matrix (default: FALSE).}

\item{checkConvergence}{Logical, whether to check for convergence of the Gibbs sampler (default: FALSE).}

\item{critVe, critVg, critVb, critPi}{Convergence criteria for residual, genetic, and marker variances, and inclusion probabilities.}

\item{ntrial}{Number of trials to assess convergence (default: 1).}

\item{verbose}{Logical, whether to print detailed output for debugging (default: FALSE).}

\item{eigen_threshold}{Threshold for eigenvalues in eigen decomposition (default: 0.995).}

\item{nit}{Number of iterations in the MCMC sampler (default: 5000).}

\item{nburn}{Number of burn-in iterations (default: 500).}

\item{nthin}{Thinning interval for MCMC (default: 5).}

\item{method}{The regression method to use, options include "blup", "bayesN", "bayesA", "bayesL", "bayesC", "bayesR".}

\item{algorithm}{Algorithm for MCMC sampling, options include "mcmc", "em-mcmc", "mcmc-eigen".}

\item{seed}{Random seed for reproducibility (default: 10).}

\item{cset_r2, r2}{R-squared thresholds for pruning (default: 0.5 and 0.05, respectively).}
}
\value{
Returns a list structure including the following components:
\item{bm}{Vector or matrix of posterior means for marker effects.}
\item{dm}{Vector or matrix of posterior means for marker inclusion probabilities.}
\item{vb}{Scalar or vector of posterior means for marker variances.}
\item{vg}{Scalar or vector of posterior means for genomic variances.}
\item{ve}{Scalar or vector of posterior means for residual variances.}
\item{rb}{Matrix of posterior means for marker correlations.}
\item{rg}{Matrix of posterior means for genomic correlations.}
\item{re}{Matrix of posterior means for residual correlations.}
\item{pi}{Vector of posterior probabilities for models.}
\item{h2}{Vector of posterior means for model probability.}
\item{param}{List of current parameters used for restarting the analysis.}
\item{stat}{Matrix of marker information and effects used for genomic risk scoring.}
}
\description{
In the Bayesian multiple regression model, the posterior density of the model parameters depends
on the likelihood of the data given the parameters and a prior probability for the model parameters.
The choice of the prior for marker effects can influence the type and extent of shrinkage induced in the model.
}
\details{
This function implements Bayesian linear regression models to provide unified mapping of 
genetic variants, estimate genetic parameters (e.g. heritability), and predict disease risk.
It is designed to handle various genetic architectures and scale efficiently with large datasets.
}
\examples{

# Plink bed/bim/fam files
bedfiles <- system.file("extdata", paste0("sample_chr",1:2,".bed"), package = "qgg")
bimfiles <- system.file("extdata", paste0("sample_chr",1:2,".bim"), package = "qgg")
famfiles <- system.file("extdata", paste0("sample_chr",1:2,".fam"), package = "qgg")

# Prepare Glist
Glist <- gprep(study="Example", bedfiles=bedfiles, bimfiles=bimfiles, famfiles=famfiles)

# Simulate phenotype
sim <- gsim(Glist=Glist, chr=1, nt=1)

# Compute single marker summary statistics
stat <- glma(y=sim$y, Glist=Glist, scale=FALSE)
str(stat)

# Define fine-mapping regions 
sets <- Glist$rsids
Glist$chr[[1]] <- gsub("21","1",Glist$chr[[1]]) 
Glist$chr[[2]] <- gsub("22","2",Glist$chr[[2]]) 

# Fine map
fit <- gmap(Glist=Glist, stat=stat, sets=sets, verbose=FALSE, 
            method="bayesR", algorithm="mcmc-eigen",nit=1500, nburn=500, nthin=5)
            
fit$post  # Posterior inference for every fine-mapped region
fit$conv  # Convergence statistics for every fine-mapped region

# Posterior inference for marker effect
head(fit$stat)             

}
\author{
Peter Sørensen
}
