% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_genotype_matrix.R
\name{gprep}
\alias{gprep}
\title{Prepare genotype data for all statistical analyses (initial step)}
\usage{
gprep(study = NULL, fnRAW = NULL, bedfiles = NULL, bimfiles = NULL,
  famfiles = NULL, ids = NULL, rsids = NULL, overwrite = FALSE,
  ncores = 1)
}
\arguments{
\item{study}{name of the study}

\item{fnRAW}{name of binary file used for storing genotypes on disk}

\item{bedfiles}{a vector names for the bed files}

\item{bimfiles}{a vector names for the bim files}

\item{famfiles}{a vector names for the fam files}

\item{ids}{individual IDs used in study}

\item{rsids}{marker rsids used in study}

\item{overwrite}{logical if TRUE overwite binary genotype file}

\item{ncores}{used to process genotypes}
}
\value{
Returns a list structure with information about genotypes
}
\description{
The gprep function prepares a socalled Glist structure providing information about the 
genotype data used in downstream analyses.

The genotypes are allele counts of the alternative allele stored as a matrix in a 
binary file (PLINK bedfile format, memory usage = (n x m)/4 bytes, where n is the number of 
individuals and m is the number of genotypes markers). 

The Glist structure provides an infrastructure for efficient processing of large-scale genetic 
data and is used as input parameter for a number of qgg core functions for 1) fitting linear 
mixed models (gsolve), 2) construction and eigen-value decomposition of genomic relationship 
matrices (grm), 3) estimating genetic parameters such heritability and correlation (greml), 
4) genomic prediction and genetic risk profiling (gsolve and gscore), and 5) single or 
multi-marker association analyses (lma and gsea).

Glist is typically only prepared once and then saved in an Rdata file. It can then at a later 
stage be loaded into R and used as input parameter in the qgg functions fascilitating 
batch processing of genotypes without having to load all genotypes into memory.
}
\examples{
#Glist <- gprep( bedfiles, bimfiles, study, path, additional arguments...)
#W <- getW( Glist, ids, rsids, additional arguments...)
}
\author{
Peter Soerensen
}
