% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_grm.R
\name{grm}
\alias{grm}
\title{Computation or eigen value decompostion of a genomic relationship matrix (GRM)}
\usage{
grm(Glist = NULL, GRMlist = NULL, ids = NULL, rsids = NULL,
  rws = NULL, cls = NULL, W = NULL, method = "add", scaled = TRUE,
  msize = 100, ncores = 1, fnG = NULL, overwrite = FALSE,
  returnGRM = FALSE, miss = 0, task = "grm")
}
\arguments{
\item{Glist}{a list providing information about genotypes stored on disk}

\item{ids}{individual ids used for computing GRM}

\item{rsids}{marker rsids used for computing GRM}

\item{rws}{rows in genotype matrix used for computing GRM}

\item{cls}{columns in genotype matrix used for computing GRM}

\item{W}{matrix of centered and scaled genotypes}

\item{method}{used for computing GRM including additive, dominance or epistasis}

\item{msize}{number of markers used for block processing of genotypes for large data}

\item{fnG}{name of the binary file used for storing the GRM on disk}

\item{overwrite}{logical if TRUE the binary file fnG will be overwritten}

\item{returnGRM}{logical if TRUE function returns the GRM matrix}

\item{miss}{is the missing code used for missing values in the genotype data}

\item{task}{computation (task="grm") or eigen value decompostion (task="eigen")}
}
\value{
Returns a GRM if returnGRM=TRUE or else a list structure (GRMlist) with information about the GRM  stored on disk
}
\description{
A function for computing or eigen value decompostion of a genomic relationship matrix 

The output of the computeGRM function is a GRMlist structure containing information 
about the genotypes stored in a binary file on disk. The Glist structure (created
only once) is used in GREML and eigen value analyses and should be saved in Rdata file
}
\examples{

# Simulate data
W <- matrix(rnorm(20000000), ncol = 10000)
	colnames(W) <- as.character(1:ncol(W))
	rownames(W) <- as.character(1:nrow(W))

# Compute GRM
GRM <- grm(W = W)

# Eigen value decompostion GRM
eig <- grm(GRM=GRM, task="eigen")
}
\author{
Peter Soerensen
}
