% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_marker_set_test.R
\name{gsea}
\alias{gsea}
\title{Gene set enrichment analysis}
\usage{
gsea(stat = NULL, sets = NULL, threshold = 0.05, method = "sum",
  nperm = 1000, ncores = 1)
}
\arguments{
\item{stat}{vector or matrix of single marker statistics (e.g. marker effects, t-stat, p-values)}

\item{sets}{list of marker sets - names corresponds to rownames in stat}

\item{threshold}{used if method='hyperG' (threshold=0.05 is deafult)}

\item{method}{including sum, cvat, hyperG, score}

\item{nperm}{number of permutations}

\item{ncores}{number of cores}

\item{W}{matrix of centered and scaled genotypes (used if method = cvat or score)}

\item{fit}{is the fit object obtained from a linear mixed model fit using the greml function}

\item{g}{vector (or matrix) of genetic effects obtained from a linear mixed model fit (GBLUP of GFBLUP)}

\item{s}{vector (or list) of single marker effects obtained from a linear mixed model fit (GBLUP of GFBLUP)}
}
\value{
Returns a dataframe or a list including 
\item{stat}{marker set test statistics} 
\item{m}{number of markers in the set}
\item{p}{p-value for marker set}
}
\description{
Gene set enrichment analyses (i.e. genetic marker set tests) using a range of methods.

The general procedure is to obtain single marker effects, from which it is possible to compute and 
evaluate a test statistic for a set of genetic markers, measuring the degree of association 
between the marker set and the phenotype. This includes the statistical model 
and the underlying assumptions, test statistics for the set of genetic markers, 
and statistical procedures for assessing the statistical significance of the observed 
test statistic under a specific null hypothesis.
}
\details{
The sum test is based on summing the single genetic marker test statistics.
The single marker test statistics can be obtained from GBLUP and GFBLUP model fits or from single marker association models such mlma or lma. 
The sum test is powerful if the genomic feature harbors many genetic markers having small to moderate effects. 
distributions, and an empirical distribution is required.
Genetic marker set tests based on the covariance statistics for a set of genetic markers.
The covariance test statistic is derived from a GBLUP (or GFBLUP) model fit. It is a measure of covariance between the total genomic effect for all markers 
and the genomic effect for the genetic markers in the genomic feature. It also relates to the explained sums of
squares for the genetic markers. 
The distribution of this test statistic under the null hypothesis (associated markers are picked at random from the total 
number of tested genetic markers) is difficult to describe in terms of exact or approximate 
Genetic marker set tests based on the hyperG test statistics for a set of genetic markers.
The hyperG marker set test tests a predefined set of markers (i.e. those within a particular genomic feature)
for an association with the trait phenotype.
Under the null hypothesis (associated markers are picked at random from the total number of tested 
genetic markers) it is assumed that the observed count statistic is a realization from a hypergeometric 
distribution.
}
\examples{

 
 # Simulate data
 W <- matrix(rnorm(20000000), ncol = 10000)
 colnames(W) <- as.character(1:ncol(W))
 rownames(W) <- as.character(1:nrow(W))
 y <- rowSums(W[, 1:10]) + rowSums(W[, 1001:1010]) + rnorm(nrow(W))
 
 # Create model
 data <- data.frame(y = y, mu = 1)
 fm <- y ~ 0 + mu
 X <- model.matrix(fm, data = data)
 
 # Single marker association analyses
 ma <- lma(y=y,X=X,W=W)
 
 # Create marker sets
 f <- factor(rep(1:100,each=100), levels=1:100) 
 sets <- split(as.character(1:10000),f=f) 

 # Set test based on sums 
 mma <- gsea(stat = ma[,"stat"]**2, sets = sets, method = "sum", nperm = 10000)
 head(mma)
 
 # Set test based on hyperG 
 mma <- gsea(stat = ma[,"p"], sets = sets, method = "hyperg", threshold = 0.05)
 head(mma)
 
 
}
\author{
Peter Soerensen
}
