% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_simulation.R
\name{gsim}
\alias{gsim}
\title{Genomic simulation}
\usage{
gsim(nt = 1, W = NULL, n = 1000, m = 1000)
}
\arguments{
\item{nt}{number of traits}

\item{W}{matrix of centered and scaled genotypes}

\item{n}{number of individuals}

\item{m}{number of markers}
}
\description{
The gsolve function is used for solving of linear mixed model equations. The algorithm used to solve the equation
system is based on a Gauss-Seidel (GS) method (matrix-free with residual updates) that handles large data sets.

The linear mixed model fitted can account for multiple traits, multiple genetic factors (fixed or random genetic
marker effects), adjust for complex family relationships or population stratification, and adjust for other
non-genetic factors including lifestyle characteristics. Different genetic architectures (infinitesimal,
few large and many small effects) is accounted for by modeling genetic markers in different sets as fixed or
random effects and by specifying individual genetic marker weights.
}
\examples{

# Simulate data
W <- matrix(rnorm(1000000), ncol = 1000)
	colnames(W) <- as.character(1:ncol(W))
	rownames(W) <- as.character(1:nrow(W))
m <- ncol(W)
causal <- sample(1:ncol(W),50)
y <- rowSums(W[,causal]) + rnorm(nrow(W),sd=sqrt(50))

X <- model.matrix(y~1)

Sg <- 50
Se <- 50
h2 <- Sg/(Sg+Se)
lambda <- Se/(Sg/m)
lambda <- m*(1-h2)/h2

# BLUP of single marker effects and total genomic effects based on Gauss-Seidel procedure
fit <- gsolve( y=y, X=X, W=W, lambda=lambda)


}
\author{
Peter Soerensen
}
