% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{hwe}
\alias{hwe}
\title{Perform Hardy Weinberg Equilibrium Test}
\usage{
hwe(Glist = NULL)
}
\arguments{
\item{Glist}{is a list structure with information about genotypes stored on disk}
}
\description{
Perform Hardy Weinberg Equilibrium Test ££(HWE)
}
\keyword{internal}
