% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_grm.R
\name{mergeGRM}
\alias{mergeGRM}
\title{Merge multiple GRMlist objects}
\usage{
mergeGRM(GRMlist = NULL)
}
\arguments{
\item{GRMlist}{list providing information about GRM matrix stored in binary files on disk}
}
\description{
Merge multiple GRMlist objects each with information about a 
genomic rfelationship matrix stored on disk
}
