% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{plotForest}
\alias{plotForest}
\title{Forest plot}
\usage{
plotForest(
  x = NULL,
  sd = NULL,
  cex = 1,
  mar = NULL,
  mai = NULL,
  xlim = NULL,
  pos = NULL,
  reorder = TRUE,
  xaxis = TRUE,
  main = NULL,
  xlab = "x"
)
}
\arguments{
\item{x}{is a vector of values}

\item{sd}{is a vector of sd values}
}
\description{
Forest plot
}
\keyword{internal}
