% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_matrix.R
\name{plotLD}
\alias{plotLD}
\title{Plot LD Matrix}
\usage{
plotLD(LD = NULL, cols = NULL)
}
\arguments{
\item{LD}{A matrix representing the LD values to be plotted. Each element should be 
a numeric value, typically between 0 and 1, representing the degree of LD.
Rows and columns of the matrix should correspond to specific genetic markers (e.g., SNPs).}

\item{cols}{A color palette to use for the plot. By default, it creates a blue gradient 
ranging from light blue ('#f0f3ff') to dark blue ('#0033BB').}
}
\value{
A plot visualizing the LD matrix. Row and column names of the LD matrix are used 
        as labels on the x and y axes, respectively.
}
\description{
Visualizes the linkage disequilibrium (LD) matrix using a color gradient.
The function produces an image plot with custom color mapping.
}
\keyword{internal}
