% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_score.R
\name{plotROC}
\alias{plotROC}
\title{Plot Receiver Operating Curves}
\usage{
plotROC(roc.data = NULL, cols = NULL)
}
\arguments{
\item{roc.data}{data frame with ROC information (from computeROC)}

\item{cols}{which columns should be used in the ROC plot}
}
\description{
Plot ROC
}
\author{
Palle Duun Rohde
}
\keyword{internal}
