% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{predict_auc_mt_continous}
\alias{predict_auc_mt_continous}
\title{Expected AUC for prediction of a binary trait using information on a correlated continuous trait}
\usage{
predict_auc_mt_continous(h2x, Nx, Kx, Px, h2y, Ny, rg, Me, M)
}
\arguments{
\item{h2x}{is the heritability of the target trait}

\item{Nx}{is the number of samples for the target trait}

\item{Kx}{is the prevalence of the target trait}

\item{Px}{is the case-control proportion of the target trait}

\item{h2y}{is the heritability of the correlated trait}

\item{Ny}{is the number of samples for the correlated trait}

\item{rg}{is the genetic correlation between the target and correlated trait}

\item{Me}{is the number of independent chromosome segments}

\item{M}{is the number of markers}

\item{Ky}{is the prevalence of the correlated trait}

\item{Py}{is the case-control proportion of the correlated trait}
}
\description{
Expected AUC for prediction of a binary trait using information on a correlated continuous trait
}
\keyword{internal}
