% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{predict_r2_mt}
\alias{predict_r2_mt}
\title{Expected R2 for multiple trait prediction of continuous traits}
\usage{
predict_r2_mt(h2x, Nx, h2y, Ny, rg, Me, M)
}
\arguments{
\item{h2x}{is the heritability of the target trait}

\item{Nx}{is the number of samples for the target trait}

\item{h2y}{is the heritability of the correlated trait}

\item{Ny}{is the number of samples for the correlated trait}

\item{rg}{is the genetic correlation between the target and correlated trait}

\item{Me}{is the number of independent chromosome segments}

\item{M}{is the number of markers}
}
\description{
Expected R2 for multiple trait prediction of continuous traits
}
\keyword{internal}
