% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{predict_r2_st}
\alias{predict_r2_st}
\title{Expected R2 for single trait prediction of a continuous trait}
\usage{
predict_r2_st(h2x, Nx, Me, M)
}
\arguments{
\item{h2x}{is the heritability of the target trait}

\item{Nx}{is the number of samples for the target trait}

\item{Me}{is the number of independent chromosome segments}

\item{M}{is the number of markers}
}
\description{
Expected R2 for single trait prediction of a continuous trait ££ continuous spelled wrong in code (line 182)
}
\keyword{internal}
