% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_genotype_matrix.R
\name{prepG}
\alias{prepG}
\title{Prepare genotype data for all statistical analyses (initial step)}
\usage{
prepG(study = NULL, fnRAW = NULL, bedfiles = NULL, bimfiles = NULL,
  famfiles = NULL, ids = NULL, rsids = NULL, overwrite = TRUE,
  ncores = 1)
}
\arguments{
\item{study}{name of the study}

\item{fnRAW}{name of binary file used for storing genotypes on disk}

\item{bedfiles}{a vector names for the bed files}

\item{bimfiles}{a vector names for the bim files}

\item{famfiles}{a vector names for the fam files}

\item{ids}{individual IDs used in study}

\item{rsids}{marker rsids used in study}

\item{overwrite}{logical if TRUE overwite binary genotype file}

\item{ncores}{used to process genotypes}
}
\value{
Returns a list structure with information about genotypes
}
\description{
The processed genotypes are stored in a matrix W. 
By default genotypes are allele counts of alternative allele which are centered and scaled. 

Currently the raw genotype format supported is plink bed/bim/fam files. 
To fascilitate a simple and common interface for using raw genotypes we prepare a Glist structure 
which contains information about the raw and processed genotypes. 
Glist is obtained using the prepG function based on plink bed/bim/fam files.
The genotype matrix, W, is too big to fit into memory and is therefore stored in a binary file.
By default the W matrix is stored in a binary file (n*m*4 bytes).
Information about the processed genotype matrix W is provided in a Glist structure. 

The output of the prepG functions is a Glist structure containing information 
about the genotype matrix W used in downstream analyses. This should be saved in Rdata file and used 
for downstream analyses. Most users do not use this directly. Furthermore this structure 
also allow us to change the underlying data structures without apparent changes in 
the Glist interface.
}
\examples{
#Glist <- prepG( bedfiles, bimfiles, study, path, additional arguments...)
#W <- getW( Glist, ids, rsids, additional arguments...)
}
\author{
Peter Soerensen

Peter S?rensen
}
