% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{rnag}
\alias{rnag}
\title{Compute Nagelkerke R2}
\usage{
rnag(yobs = NULL, ypred = NULL)
}
\arguments{
\item{yobs}{is a vector of observed phenotypes}

\item{ypred}{is a vector of predicted phenotypes}
}
\description{
Compute Nagelkerke R2 ££ perhaps add: r2nag?
}
\keyword{internal}
