% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_bayes.R
\name{splitWithOverlap}
\alias{splitWithOverlap}
\title{Split Vector with Overlapping Segments}
\usage{
splitWithOverlap(vec, seg.length, overlap)
}
\arguments{
\item{vec}{A numeric or character vector to be split.}

\item{seg.length}{An integer specifying the length of each segment.}

\item{overlap}{An integer specifying the number of overlapping elements between consecutive segments.}
}
\value{
A list where each element is a segment of the input vector. The segments can overlap based on the specified overlap.
}
\description{
Splits a vector into segments of a specified length with a specified overlap.
The function returns a list where each element contains a segment of the vector.
}
\keyword{internal}
