% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qgg.R
\name{aireml}
\alias{aireml}
\title{Genomic Feature Model analyses implemented using Restriced Likelihood Methods in DMU}
\usage{
aireml(fm = NULL, vfm = NULL, Klist = NULL, restrict = NULL,
  data = NULL, validate = NULL, bin = NULL)
}
\arguments{
\item{fm}{a formula with model statement for the linear mixed model}

\item{Klist}{a list of relationship/correlation matrices corresponding to random effects specified in vfm}

\item{data}{a data frame containing the phenotypic observations and fixed factors specified in the model statements}

\item{validate}{a matrix or a list with the ids of validation sets corresponding to the rows in data}

\item{bin}{is the directory for DMU binaries (dmu1 and dmuai1)}
}
\value{
Returns results in a list structure including
\item{f}{list of predicted random effects}
\item{sigma}{estimated variance components}
\item{asd}{asymptotic standard deviation for the estimated variance components}
\item{fitted}{ fitted values from linear mixed model fit}
\item{residuals}{ residuals from linear mixed model fit}
}
\description{
Genomic Feature Best Linear Unbiased Prediction models implemented using REML.
}
\details{
The models are implemented using restricted maximum likelihood methods.
Variance components estimated using REML and predictions are based on MME.
Predicted random effects and single marker effects and statistics can be obtained.
Cross validation procedures for assessment of prediction accuracy and model selection.
This is an interface to be used for DMU.
}
\examples{
library(qgg)

setwd("C:/Users/sor/Dropbox/GFBLUP DGRP/scripts used in the analyses/work")
#bin <- "C:/Program Files (x86)/QGG-AU/DMUv6/R5.2-EM64T/bin"

# Simulate data
W <- matrix(rnorm(4000000),ncol=10000)
  colnames(W) <- as.character(1:ncol(W))
  rownames(W) <- as.character(1:nrow(W))

G <- W\%*\%t(W)/ncol(W)

y <- rowSums(W[,1:10]) + rowSums(W[,1001:1010]) + 10*rnorm(nrow(W))

data <- data.frame(f=factor(sample(1:2,nrow(W),replace=TRUE)),g=factor(1:nrow(W)),y=y )

fm <- y ~ f +  (1|g~G)
Klist <- list( G=G )


fit <- aireml(fm=fm,Klist=Klist,data=data)
  str(fit)
}
\author{
Peter Sørensen
}
\references{
Mapping Variants to Gene Ontology Categories Improves Genomic Prediction for Quantitative Traits in Drosophila melanogaster. Under review Genetics (2016). Edwards SM, Sørensen IF, Sarup P, Mackay TF, Sørensen P.

Genomic BLUP Derived Set Tests Identify Genetic Variants Associated with Schizophrenia in Functionally Associated Biological Processes. Under review, Genetics (2015). Rohde PD, Demontis D, The GEMS Group, Børglum AD, Sørensen P.

Partitioning of genomic variance reveals biologic pathways associated with udder health and milk production traits in dairy cattle. GSE (2015) 47:60. Edwards SM, Thomsen B, Madsen P, Sørensen P.

Increased prediction accuracy using a genomic feature model including prior information on quantitative trait locus regions in purebred Danish Duroc pigs. BMC Genetics (2016) 17:11. Sarup P, Jensen J, OstersenT, Henryon M, Sørensen P.
}

