% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_grm.R
\name{computeGRM}
\alias{computeGRM}
\title{Compute genomic relationship matrix (GRM)}
\usage{
computeGRM(Glist = NULL, ids = NULL, rsids = NULL, rws = NULL,
  cls = NULL, W = NULL, method = "add", scaled = TRUE,
  msize = 100, ncores = 1, fnG = NULL, overwrite = FALSE,
  returnGRM = FALSE, miss = 0)
}
\arguments{
\item{Glist}{information about genotypes}

\item{ids}{individual ids used for computing GRM}

\item{rsids}{marker rsids used for computing GRM}

\item{rws}{rows in genotype matrix used for computing GRM}

\item{cls}{columns in genotype matrix used for computing GRM}

\item{W}{matrix of centered and scaled genotypes}

\item{method}{used for computing GRM including additive, dominance or epistasis}

\item{scaled}{should genotypes be scaled (default: scaled=TRUE)}

\item{msize}{number of markers used for block updates}

\item{fnG}{name of file for storing GRM on disk}

\item{overwrite}{logical should the file fnG be overwritten}

\item{returnGRM}{should the function return the GRM matrix}

\item{miss}{missing value code used in genotype matrix}
}
\value{
Returns GRM if returnGRM=TRUE or else a list structure (GRMlist) with information about the GRM  stored on disk
}
\description{
A function for computing a genomic relationship matrix. 

The output of the prepG and computeRAW functions is a Glist structure containing information
}
\details{
Individuals may be subsetted for additional analyses such as cross validation.
}
