% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_grm.R
\name{computeGRM}
\alias{computeGRM}
\title{Compute genomic relationship matrix (GRM)}
\usage{
computeGRM(Glist = NULL, ids = NULL, rsids = NULL, rws = NULL,
  cls = NULL, W = NULL, method = "add", scaled = TRUE, msize = 100,
  ncores = 1, fnG = NULL, overwrite = FALSE, returnGRM = FALSE,
  miss = 0)
}
\arguments{
\item{Glist}{a list providing information about genotypes stored on disk}

\item{ids}{individual ids used for computing GRM}

\item{rsids}{marker rsids used for computing GRM}

\item{rws}{rows in genotype matrix used for computing GRM}

\item{cls}{columns in genotype matrix used for computing GRM}

\item{W}{matrix of centered and scaled genotypes}

\item{method}{used for computing GRM including additive, dominance or epistasis}

\item{msize}{number of markers used for block processing of genotypes for large data}

\item{fnG}{name of the binary file used for storing the GRM on disk}

\item{overwrite}{logical if TRUE the binary file fnG will be overwritten}

\item{returnGRM}{logical if TRUE function returns the GRM matrix}

\item{miss}{is the missing code used for missing values in the genotype data}
}
\value{
Returns a GRM if returnGRM=TRUE or else a list structure (GRMlist) with information about the GRM  stored on disk
}
\description{
A function for computing a genomic relationship matrix. 

The output of the computeGRM function is a GRMlist structure containing information 
about the genotypes stored in a binary file on disk. The Glist structure (created
only once) is used in GREML and eigen value analyses and should be saved in Rdata file
}
\author{
Peter Soerensen
}
