% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_matrix.R
\name{getLD}
\alias{getLD}
\title{Get sparse LD matrix}
\usage{
getLD(Glist = NULL, chr = NULL, rsids = NULL)
}
\arguments{
\item{Glist}{list structure with information about genotypes stored on disk}

\item{chr}{chromosome from which LD sets are extracted}

\item{rsids}{included in the sparse LD matrix extracted}
}
\description{
Get sparse LD matrix in Glist
}
\keyword{internal}
