% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_matrix.R
\name{getLDsets}
\alias{getLDsets}
\title{Get marker LD sets}
\usage{
getLDsets(Glist = NULL, chr = NULL, r2 = 0.5)
}
\arguments{
\item{Glist}{list structure with information about genotypes stored on disk}

\item{chr}{chromosome for which LD sets are extracted}

\item{r2}{threshold used for extracting LD sets}
}
\description{
Get marker LD sets based on sparse LD matrix in Glist
}
\keyword{internal}
