% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_matrix.R
\name{getMarkers}
\alias{getMarkers}
\title{Get marker rsids in a genome region}
\usage{
getMarkers(Glist = NULL, chr = NULL, region = NULL)
}
\arguments{
\item{Glist}{list structure with information about genotypes stored on disk}

\item{chr}{chromosome from which markers are extracted}

\item{region}{genome region (in base pairs) from which markers are extracted}
}
\description{
Get marker rsids in a specific genome region based on marker in Glist
}
\keyword{internal}
