% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_genotype_matrix.R
\name{gprep}
\alias{gprep}
\title{Prepare genotype data for all statistical analyses (initial step)}
\usage{
gprep(study = NULL, fnRAW = NULL, bedfiles = NULL, bimfiles = NULL,
  famfiles = NULL, ids = NULL, rsids = NULL, overwrite = FALSE,
  ncores = 1)
}
\arguments{
\item{study}{name of the study}

\item{fnRAW}{name of binary file used for storing genotypes on disk}

\item{bedfiles}{a vector names for the bed files}

\item{bimfiles}{a vector names for the bim files}

\item{famfiles}{a vector names for the fam files}

\item{ids}{individual IDs used in study}

\item{rsids}{marker rsids used in study}

\item{overwrite}{logical if TRUE overwite binary genotype file}

\item{ncores}{used to process genotypes}
}
\value{
Returns a list structure with information about genotypes
}
\description{
To fascilitate a simple and common interface for using large-scale genotype data we prepare 
a socalled Glist structure providing information about the raw and processed genotypes. 
The Glist structure is obtained using the gpred function based on plink bed/bim/fam files. 
The Glist should be saved in an Rdata file and is used as input to several functions 
(e.g. greml,grm,gsea,gscore,gsolve,lma) used for the statistical analyses implemented in qgg.

The genotypes are allele counts of the alternative allele based on the PLINK bed file format. 
Genotype data (i.e. genotype matrix) are stored in a binary file (nxm/4 bytes, 
n is the number of individuals and m is the number of genotypes markers).
}
\examples{
#Glist <- gprep( bedfiles, bimfiles, study, path, additional arguments...)
#W <- getW( Glist, ids, rsids, additional arguments...)
}
\author{
Peter Soerensen
}
