% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_genotype_matrix.R
\name{gprep}
\alias{gprep}
\title{Prepare genotype data for all statistical analyses (initial step)}
\usage{
gprep(study = NULL, fnRAW = NULL, bedfiles = NULL, bimfiles = NULL,
  famfiles = NULL, ids = NULL, rsids = NULL, overwrite = FALSE,
  ncores = 1)
}
\arguments{
\item{study}{name of the study}

\item{fnRAW}{path and .raw filename of the binary file used for storing genotypes on the disk}

\item{bedfiles}{a vector of names for the PLINK bed-files}

\item{bimfiles}{a vector of names for the PLINK bim-files}

\item{famfiles}{a vector of names for the PLINK fam-files}

\item{ids}{a vector of individual IDs used in the study}

\item{rsids}{a vector of genotype marker rsids used in the study}

\item{overwrite}{logical if TRUE overwite binary genotype file}

\item{ncores}{number of cores used to process the genotypes}
}
\value{
Returns a list structure with information about genotypes
}
\description{
All functions in qgg relies on a simple infrastructure that takes four main input sources; 
phenotype data (y), covariate data (X), a Glist object that contains information about the 
genotype data (for smaller data sets that can be in memory, the Glist object is not required, 
and genotypes can be stored in a n(individuals) x m(SNP markers) matrix of centered and scaled 
genotypes (G), and a genomic relationship matrix (GRM).

The gprep function prepares the Glist, and is required for downstream analyses of large-scale 
genetic data. Typically, the Glist is prepared once, and saved as an *.Rdata-file. The Glist 
structure is used as input parameter for a number of qgg core functions for; 1) construction 
of genomic relationship matrices (grm), 2) estimating genetic parameters (greml), 
3) single SNP association approximations (lma), 4) gene set enrichment analyses (gsea), 
and 5) genomic prediction from genotypes (gsolve) or SNP summary statistics (gscore).

The gprep function reads genotype information from binary PLINK files, and creates the Glist 
object that contains general information about the genotypes such as reference alleles, 
allele frequencies and missing genotypes, and construct a binary file on the disk that contains 
the genotypes as allele counts of the alternative allele (memory usage = (n x m)/4 bytes).
}
\examples{
#Glist <- gprep( bedfiles, bimfiles, study, path, additional arguments...)
#W <- getW( Glist, ids, rsids, additional arguments...)
}
\author{
Peter Soerensen
}
