% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_solve.R
\name{gscore}
\alias{gscore}
\title{Genomic prediction based on single marker summary statistics}
\usage{
gscore(Glist = NULL, S = NULL, ids = NULL, rsids = NULL,
  rws = NULL, cls = NULL, scaled = TRUE, msize = 100, ncores = 1)
}
\arguments{
\item{Glist}{list of information about genotype matrix}

\item{S}{matrix of single marker effects}

\item{ids}{vector of individuals used in the analysis}

\item{rsids}{vector marker rsids used in the analysis}

\item{ncores}{number of cores}
}
\description{
Function for genomic predictions based on single marker summary statistics and observed phenotypes
}
