% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_solve.R
\name{gsolve}
\alias{gsolve}
\title{Genomic prediction based on fitting a linear mixed model}
\usage{
gsolve(y = NULL, X = NULL, Glist = NULL, ids = NULL, rsids = NULL,
  sets = NULL, scaled = TRUE, lambda = NULL, weights = FALSE,
  maxit = 500, tol = 1e-05, ncores = 1)
}
\arguments{
\item{y}{vector or matrix of phenotypes}

\item{X}{design matrix of fixed effects}

\item{Glist}{list of information about genotype matrix}

\item{ids}{vector of individuals used in the analysis}

\item{rsids}{vector marker rsids used in the analysis}

\item{lambda}{vector of single marker weights used in BLUP}

\item{maxit}{maximum number of iterations of reml analysis}

\item{tol}{tolerance, i.e. the maximum allowed difference between two consecutive iterations of reml to declare convergence}

\item{ncores}{number of cores}

\item{W}{matrix of centered and scaled genotypes}
}
\description{
Genomic predictions based on solving linear mixed model equations.

The linear mixed model fitted can account for multiple genetic factors (fixed or random genetic marker effects), 
adjust for complex family relationships or population stratification, and adjust for other non-genetic factors 
including lifestyle characteristics. The linear mixed model can easily be extended to fit multiple correlated phenotypes. 
Different genetic architectures (infinitesimal, few large and many small effects) is accounted for by modeling 
genetic markers in different sets as fixed or random effects and by specifying individual genetic marker weights. 
Different genetic models (e.g. additive and non-additive) can be specified.
}
\author{
Peter Soerensen
}
