% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_marker_set_test.R
\name{mma}
\alias{mma}
\title{Genetic marker set tests}
\usage{
mma(stat = NULL, sets = NULL, ncores = 1, nperm = 1000,
  method = "sum")
}
\arguments{
\item{stat}{vector of single marker statistics (e.g. marker effects, t-stat, p-values)}

\item{sets}{list of marker sets - names corresponds to rownames in stat}

\item{ncores}{number of cores}

\item{nperm}{number of permutations}

\item{method}{including sum, cvat, hyperG, score}

\item{W}{matrix of centered and scaled genotypes (used if method = cvat or score)}

\item{fit}{is the fit object obtained from a linear mixed model fit using the greml function}

\item{g}{vector (or list) of genetic effects obtained from a linear mixed model fit (GBLUP of GFBLUP)}

\item{s}{vector (or list) of single marker effects obtained from a linear mixed model fit (GBLUP of GFBLUP)}

\item{threshold}{used if method = hyperG}
}
\value{
Returns a dataframe including 
\item{setT}{marker set test statistics} 
\item{nset}{number of markers in the set}
\item{p}{p-value for marker set}
}
\description{
Genetic marker set tests based on summing the single genetic marker test statistics.
The sum test is powerful if the genomic feature harbors many genetic markers having small to moderate effects. 
Genetic marker set tests based on the covariance statistics for a set of genetic markers.
The covariance test statistic is derived from a GBLUP (or GFBLUP) model fit. It is a measure of covariance between the total genomic effect for all markers 
and the genomic effect for the genetic markers in the genomic feature. It also relates to the explained sums of
squares for the genetic markers. 
The distribution of these test statistics under the null hypothesis is difficult to describe in terms of exact or approximate 
distributions, and an empirical distribution is required.
Genetic marker set tests based on the hyperG test statistics for a set of genetic markers.
}
\details{
The single marker test statistics can be obtained from GBLUP and GFBLUP model fits or from standard GWAS. 
The distribution of this test statistic under the null hypothesis (associated markers are picked at random from the total 
number of tested genetic markers) is difficult to describe in terms of exact or approximate 
distributions, and an empirical distribution is required.
The hyperG marker set test tests a predefined set of markers (i.e. those within a particular genomic feature)
for an association with the trait phenotype.
Under the null hypothesis (associated markers are picked at random from the total number of tested 
genetic markers) it is assumed that the observed count statistic is a realization from a hypergeometric 
distribution.
This hypothesis can be formulated and tested in a number of ways. Here we consider a test statistic based 
on counting the number of genetic markers in the feature that are associated to trait phenotype. 
A test based on the count test statistic is likely to have high power to detect association if the genomic 
feature harbours genetic markers with large effects.
}
\examples{

# Simulate data
W <- matrix(rnorm(2000000), ncol = 10000)
  colnames(W) <- as.character(1:ncol(W))
y <- rowSums(W[, 1:10]) + rowSums(W[, 1001:1010]) + rnorm(nrow(W))

# REML analyses 
data <- data.frame(y = y, mu = 1)
fm <- y ~ mu
fit <- gfm(fm = fm, W = W, sets = list(colnames(W)), data = data)
fit$df <- 10
fit$p <- pt(fit$s / sqrt(fit$vs), df = fit$df, lower.tail = FALSE) 

sets <- list(A = as.character(1:100), B = as.character(101:1000), C = as.character(1001:5000), D = as.character(5001:10000))

# Set test based on sums 
res <- mma(stat = fit$s**2, sets = sets, method = "sum", nperm = 100)

# Set test based on cvat 
res <- mma(stat = fit$s, W = W, sets = sets, method = "cvat", nperm = 100)

# Set test based on hyperG 
res <- mma(stat = fit$p, sets = sets, method = "hyperG", threshold = 0.05)

}
\references{
Rohde, P. D., Demontis, D., Cuyabano, B. C. D., Boerglum, A. D., & Soerensen, P. (2016). Covariance Association Test (CVAT) Identifies Genetic Markers Associated with Schizophrenia in Functionally Associated Biological Processes. Genetics, 203(4), 1901-1913.

Rohde, P. D., Edwards S. M., Sarup P., Soerensen, P. (August, 2014). Gene-based Association Approach Identify Genes Across Stress Traits in Fruit Flies. Poster presented at the 10th World Congress of Genetics Applied to Livestock Production (WCGALP), Vancouver, Canada.
}
\author{
Peter Soerensen
}
