% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_bayes.R
\name{plotBayes}
\alias{plotBayes}
\title{Plot fit from gbayes}
\usage{
plotBayes(fit = NULL, causal = NULL, what = "Beta", chr = 1)
}
\arguments{
\item{fit}{object from gbayes}

\item{causal}{indices for "causal" markers}

\item{what}{character fro what to plot (e.g. "trace", "Beta", "Ve", "Vg", "Vb")}

\item{chr}{chromosome to plot}
}
\description{
Summary plots from gbayes fit
}
\keyword{internal}
