% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{predict_auc_st}
\alias{predict_auc_st}
\title{Expected AUC for prediction of a binary trait}
\usage{
predict_auc_st(h2x, Nx, Kx, Px, Me, M)
}
\arguments{
\item{h2x}{is the heritability of the target trait}

\item{Nx}{is the number of samples for the target trait}

\item{Kx}{is the prevalence of the target trait}

\item{Me}{is the number of independent chromosome segments}

\item{M}{is the number of markers}
}
\description{
Expected AUC for prediction of a binary trait
}
\keyword{internal}
