% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{predict_r2_mt}
\alias{predict_r2_mt}
\title{Expected R2 for multiple trait prediction of continuous traits}
\usage{
predict_r2_mt(h2x, Nx, h2y, Ny, rg, Me, M)
}
\arguments{
\item{h2x}{Heritability of the target trait.}

\item{Nx}{Number of samples for the target trait.}

\item{h2y}{Heritability of the correlated trait.}

\item{Ny}{Number of samples for the correlated trait.}

\item{rg}{Genetic correlation between the target and correlated trait.}

\item{Me}{Number of independent chromosome segments.}

\item{M}{Number of markers.}
}
\value{
A numeric value representing the expected R2 for the multiple trait prediction.
}
\description{
Computes the expected R2 value for the multiple trait prediction of continuous traits.
}
\keyword{internal}
