% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg_utility_functions.R
\name{predict_r2_st}
\alias{predict_r2_st}
\title{Expected R2 for single trait prediction of a continuous trait}
\usage{
predict_r2_st(h2x, Nx, Me, M)
}
\arguments{
\item{h2x}{Heritability of the target trait.}

\item{Nx}{Number of samples for the target trait.}

\item{Me}{Number of independent chromosome segments.}

\item{M}{Number of markers.}
}
\value{
A numeric value representing the expected R2 for the single trait prediction.
}
\description{
Computes the expected R2 value for the single trait prediction of a continuous trait.
}
\keyword{internal}
